chore(etfs): drop API routes superseded by /full-render and direct Prisma calls
Five per-ETF GET endpoints had no remaining callers after the previous
two commits and are now removed:

  /etfs-v1/exchange/<E>/<T>/available-sections   (replaced by direct
                                                 Prisma in EtfRelatedSections)
  /etfs-v1/exchange/<E>/<T>/financial-info       (rolled into /full-render)
  /etfs-v1/exchange/<E>/<T>/scores               (rolled into /full-render)
  /etfs-v1/exchange/<E>/<T>/similar-etfs         (rolled into /full-render)
  /etfs-v1/exchange/<E>/<T>/price-history        (rolled into /full-render)

The response types that used to live in those route files
(EtfFinancialInfoResponse, EtfScoresResponse, SimilarEtf) moved to
src/types/etf/etf-detail-response-types.ts; UI components import from
there now. PriceHistoryResponse already lived in the tickers-v1 tree —
the ETF re-export is no longer needed.

Verified before deleting:
- No fetch URLs reference these paths anywhere in the repo (UI, admin,
  lambdas, scripts).
- All remaining type imports were updated to the new types file.
- yarn compile + lint + prettier-check all clean.

Routes kept (still have callers):
- /etfs-v1/exchange/<E>/<T> (basic ETF — used for metadata fetch + by
  every subpage)
- /etfs-v1/exchange/<E>/<T>/analysis (used by all 4 category subpages)
- /etfs-v1/exchange/<E>/<T>/competition (used by /competition subpage)
- /etfs-v1/exchange/<E>/<T>/portfolio-holdings (used by /holdings subpage)
- /etfs-v1/exchange/<E>/<T>/mor-info (used by /performance-returns and
  /financial-data subpages)
- /etfs-v1/exchange/<E>/<T>/full-render (the new consolidated endpoint
  the main page reads)
