+++ insights-ui/src/components/etf-reportsv1/analysis/EtfRadarChart.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { EtfScoresResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/scores/route';
+import type { EtfScoresResponse } from '@/types/etf/etf-detail-response-types';
 import { EtfAnalysisResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/analysis/route';
 import { EtfAnalysisCategory } from '@/types/etf/etf-analysis-types';
 import { findFactorDefinition } from '@/utils/etf-analysis-reports/etf-report-input-json-utils';
